error_handlerモジュールをapplicatonControllerへ移動
